Fix Accessibility: The BoundingRectangle is incorrect when clicking the list & list item in inspect tree #3655
Issue is reproduced because we get the rectangle for the ListBox, whether the ListBox is expanded or not. Unfortunately, we cannot return an empty rectangle, it is replaced with a native rectangle. Therefore, we return the rectangle with the maximum offset to the invisible zone.

Added check for "Bounds" property of ListBox items. If the ListBox is not visible (it has an "OffScreen" state), then we return an empty rectangle.