Fix ListViewItemAccessibleObject does not update data about ListViewGroup when moved to another ListViewGroup #4839 (#4841)
The issue is reproduced because during the creation of the ListViewItemAccessibilityObject, we save the current value of the ListViewGroup and it is no longer updated. This is not the right approach because the user can change the ListViewGroup of a ListViewItem or remove ListViewGroups from the ListView. Currently we use the "OwningGroup" property, which, on each call, calculates whether the ListView has ListViewGroups and which ListViewGroup the ListViewItem is in.

Also, an issue with an attempt to get "Bounds" property when the Handle is not created has been fixed. Now, in the absence of a Handle, we immediately return an empty rectangle

Added and updated unit tests