Adding logic for announcing of column header by Narrator (#4853)
When the user hovers over the ColumnHeader, we receive an event for redrawing. Using this event, we get data about the columns header and the shift of coordinates due to scrolling. Using the calculated coordinates, we make a HITTEST API request. From this request, we get data about the column index and announce header using the this index.

Added unit tests