Fix Create separate accessibility objects for each ListView view #4748
During the separating of ListViewAccessibilityObject, two ways of separating were considered: "View" and ListViewGroup presence.

Unfortunately, it makes no sense to separate ListViewAccessibilityObject by "View", since the class contains almost no code specific to any "View".

A ListViewAccessibilityObject  has code specific to the presence or absence of ListViewGroups, but the display of a ListViewGroup depends on many factors:
"ShowGroups" property;
"VirtualMode" property;
"View" property;
"Groups" property;
Visual styles.
Keeping track of all of these factors will lead to the need to add changes that will be much more difficult to support than the current code. Therefore, it was decided to leave the current code for the ListViewAccessibilityObject.

When separating the ListViewItemAccessibleObject, we took into account how the ListViewItemAccessibleObject works with ListViewGroups and ListViewItems depending on the view.
The "ListViewItemBaseAccessibleObject" class has been added. It contains the implementation for most of the properties and methods of the "ListViewItemAccessibleObject" class. In addition, its implementation fully corresponds to the behavior of elements when the tree is in "LargeIcon" or "SmallIcon" view.

For the "List" view, an "ListViewItemListAccessibleObject" class has been added. It contains an overridden implementation for the "Bounds" and "RuntimeId" properties and the "FragmentNavigate" method. Since this view does not support ListViewGroup, it allows us to remove some unnecessary checks.

For the "Tile" mode, an "ListViewItemTileAccessibleObject" class has been added. It contains implementations for methods such as "FragmentNavigate", "GetChild", "GetChildInternal", "GetChildCount", "GetChildIndex" and "GetSubItemBounds" methods. This is necessary because this mode supports ListViewSubItems and has complex logic to work with them. In addition, we have a special method "GetChildInternal", which is necessary because inside the classes we operate with real indexes of ListViewSubItems , and not visible to the user.

For "Details" mode, the "ListViewItemDetailsAccessibleObject" class has been added. It does not have such overloaded logic for working with ListViewSubItems as the "ListViewItemTileAccessibleObject" class, but it has special functionality for working with fake accessibility objects. This is necessary because the user sees the cell and can interact with it, whether there is a ListViewSubItem for it or not.

Since ListViewSubItems do not support descendants, and all the logic for interacting with each other is taken from the ListViewItemAccessibleObjects, there is no point in separating the "ListViewSubItemAccessibleObject" of the class.