Add a new code style rule (IDE0005) Rule IDE0005 - "Remove unnecessary import" It's necessary to have VS warnings, if there are forgotten extra usings in files.
To track files with this style issues,
it's necessary to add `EnforceCodeStyleInBuild` flag to all project files.
To do that Directory.Build.props is used.
If not to add this flag, there won't be warnings when building.
There will be warnings in VS only.

This commit is broken, because there are many files with this code style issue.
It will be fixed in the next commit.