[Accessibility]MAS 1.4.11: The color contrast of the focus dot line on each selection items for LitView/CheckedListBox is below 3:1 #3059
During investigation, we found that the orange color from the issue is obtained by inverting the blue (highlighted) color. Unfortunately, orange does not contrast well with most colors. Therefore, during the discussions, it was decided to change the orange color to white. Thus, we get a black and white dotted border that has the necessary contrast with the blue (highlighted) color and the colors that the user can set.

Added additional logic to the "GetFocusPen" method to get the required border colors.
Updated parameters for "DrawFocusRectangle" method.
Added "DrawBlackWhiteFocusRectangle" method.