Fix ListView evaluates all items in VirtualMode #5033
The issue is reproduced because each time we get the Accessibility object, we calculate the current index of the ListViewItemAccessibleObject and the index to get the next / previous ListViewItem. Both operations use a loop that checks all the ListViewItem in order to find the one we want. Unfortunately, when we have 1 million ListViewItem, this approach becomes too laborious.

Fixed the logic for getting indices, now we use the element's index instead of calculating it every time