Add ApplicationConfiguration.Initialize() source generator
Add a C# source generator that emits application bootstrap as described
in https://github.com/dotnet/designs/blob/main/accepted/2021/winforms/streamline-application-bootstrap.md
for the following properties:
* ApplicationVisualStyles
* ApplicationDefaultFont
* ApplicationHighDpiMode
* ApplicationUseCompatibleTextRendering

For top level statements the bootstrap also emits code to set STA mode.

A build will fail if any of the properties contain invalid or otherwise
incorrect values.

FontConverter is not available in netstandard 2.0, to bridge the gap
FontConverter logic has been partially copied from the corresponding .NET
runtime implementation to facilitate the comparable behaviour.
The only missing gap is parsing of FontFamily (which requires GDI+ calls),
which means that if a supplied font descriptor contains an invalid font
name it will cause an app to crash. This however is a desired behaviour.