Fixed BoundingRectangle width for ComboBox and ListView that have the vertical scrollbar
Fixes #4404. Currently BoundingRectangle for ComboBox and ListView doesn't take into account a vertical scrollbar if it's present. This problem is not present in .NET Framework 4.7.2 because UIA providers for these controls was added later and we need to fix this issue to keep the old behavior. When UIA provider is not present for the control UIAutomationCore uses GetWindowRect as a fallback. So, one of the possible solutions is to use GetWindowRect in the BoundingRectangle implementation. Another solution is to return an empty rectangle from the BoundingRectangle, that will trigger the fallback code in the UIAutomationCore. Using GetWindowRect directly should be the best solution since we won't rely on the internals of UIAutomationCore.