Clean up MapWindowPoints and more PropertyGrid
MapWindowPoints was not efficient as it had two separate P/Invokes and the pinning was happening in the generated interop layer. Additionally we weren't leveraging the IHandle optimization. Lastly it was risky as we were specifying the count all over the place. Cleaned this up as we were using this in PropertyGrid.

Add more comments to `CommandsPane` and a bit more cleanup to `PropertyGrid`.