Don't run tests in official build (#5576)
The tests are reasonably flaky, and frequently need manual reruns. Other repos don't run tests in their official builds for this reason as well as time. This saves a little under 40% of the build time in passing cases (though these tests are short). Tests are still run in PR and public CI scenarios.