Got rid of a static buffer in MonthCalendar control
The native calendar control sends MCN_GETDAYSSTATE notification for application to fill in bitfields representing dates that should be displayed in bold. WParam used in this notification contains this array of months. This array is allocated on the stack by the native calendar control. However winforms code was re-allocating this array and writing the pointer to a new array into the WPARAM structure. (#5561)

I think the root cause is the documentation bug here - https://docs.microsoft.com/windows/win32/controls/mcn-getdaystate because this doc https://docs.microsoft.com/cpp/mfc/setting-the-day-state-of-a-month-calendar-control?view=msvc-160 indicates that the destination array is allocated by the caller