Removed GridEntryTextBox AO from the tree
Fixes #5656

The fix doesn't allow to create and put EditAccessibleObject to the accessibility tree,
if the TextBox control is not created,
because otherwise EditAccessibleObject doesn't get correct native properties,
and a user see some AO of not created control.

PropertyGridView EditTextBox creates on the first click on the GridView
and if there is a selected GridEntry.

(cherry picked from commit 28e4b7d4ea6a4fb21bc777b83b36e7765e44bd49)