Fix MonthCalendar not firing DateSelected when first click on last date #5854
When choosing a date in the MonthCalendar, we get 2 messages: MCN.SELCHANGE and MCN.SELECT. When choosing the minimum/maximum date for the first time, we only get MCN.SELCHANGE. As a result, the "DateSelected" event is not raised.

This is due to incorrect "(start.Ticks <_minDate.Ticks || end.Ticks <_minDate.Ticks" and "(start.Ticks > _maxDate.Ticks || end.Ticks > _maxDate.Ticks)" conditions in the "MonthCalendar.WmDateChanged" method. In addition to the date, these conditions also take into account the time, whose settings are not visible and not available to the user.

As a fix, I suggest using the "Date" for comparison rather than "Ticks". In this case, the condition works correctly for the minimum and maximum dates and the "DateSelected" event is called.

Also fixed conditions in "WmDateSelected" method.

Added MAUI tests.