Fixing missing default action for combo box
Fixes #3712. Adds the default action description and the default action itself for combo box control with a dropdown list. It can also be achieved by making LegacyIAccessiblePattern and InvokePattern not supported for the accessible object of combo box (.NET Core 3.1 behavior). This way UIAutomationCore will use its own fallback mechanism to support default actions. But it has a drawback of making default action dependent on accessible role property which isn't desirable. Another drawback is that we're depending on UIAutomationCore internals/implementation details which is also not desirable. It's better to have our own implementation which is visible to everyone and easy to debug.