Fixed missing expand-collapse pattern for property grid dropdown controls
Fixes #3672.

`PropertyDescriptorGridEntry` marks `ExpandCollapsePattern` as supported only for enumerable and expandable (in `GridEntry.IsPatternSupported`) entries. Enumerable entries are entries with the predefined set of values such as colors, booleans, etc. Expandable entries are mostly for categories. The other entries with the dropdowns are treated as `DropDownEditable` which can be checked by `NeedsDropDownButton` property. The fix is to check not only `Enumerable` but also `NeedsDropDownButton` property in `PropertyDescriptorGridEntryAccessibleObject.IsPatternSupported` for `ExpandCollapsePattern`. `ExpandCollapseState` is already correctly treated by using `DropDownVisible` property. `Expand` and `Collapse` methods also correctly perform the corresponding actions. This fixes `ExpandCollapsePattern` for `MultilineStringEditor` and `DateTimeEditor`.