Fixing moving focus backward with Shift-Tab at ToolStrip with TabStop=true.
If we don't found any siblings, and the control is a ToolStripItem that hosts a control itself, then we shouldn't return its parent, because it would be the same ToolStrip we're currently at. Instead, we should return the control that is previous to the current ToolStrip.

Debug.Asserts from 6.0 are replaced for Exceptions with localized strings

(cherry picked from commit 0e53ce7b6568af6d81518616c3690807042d0dbe)