Removed nullability from RuntimeId return type (#5638)
RuntimeId cannot be null, accessibility tools expect that it is non-empty array. Added throwing of `NotSupportedException` to `AccessibleObject.RuntimeId` in order to indicate that a valid value cannot be provided. This exception is visible to the users that are accessing `RuntimeId` through `IAccessibleEx` or `IRawElementProviderFragment` interface.

`Debug.Fail` in `AccessibleObject.RuntimeId` allows developers to catch the overrides that are using `AccessibleObject.RuntimeId` during the debugging. It's not visible to the end users as debug dlls aren't provided in the sdk. As well unit-test was added to verify that all derived classes override `RuntimeId`.