Remove probing for availability of oleacc.dll #6064
Some code in the setter of the Handle property could be removed because of now we can reference https://docs.microsoft.com/en-us/windows/win32/api/oleacc/nf-oleacc-createstdaccessibleobject with no probing for oleacc.dll availability because it is available on Windows 2K+ and Server 2003+ and we not supported below these versions.

The unnecessary code was removed.