Combine PropertyGrid tab information.
Combine tab related data into a record for clarity and safety. Use a List<> instead of an array to avoid complicated insert logic.

Maintain tab selection as an instance instead of an index.

Move error handling to CreateTab so thrown exceptions with messages (missing a bitmap / name for the tab) are actually thrown instead of a silent failure.