Resume invocation of OnParentHandleRecreated for child control when form recreates handle
Fixes #6093

Fix the regression introduced in #2262, which added a check to the method
handling recreation of a control handle to verify that the control is
created before any child controls can be re-parented on to it.

However it was missed that DestroyHandle() method resets the state (i.e.
removes Created state), but CreateHandle() method doesn't set the state
back.