Added prevention of label editing when clicking on subitems of ListView
Fixes #4816. ListView allows editing the label by clicking on any subitem of a row. On the other hand, Windows Explorer allows label editing only by clicking on the first subitem (which is actually a label) of a row. Windows Explorer behavior should be more appropriate for a common user. The logic of the label editing is located in ComCtl32.dll which is provided by Windows and therefore can't be changed on our side. In order to change the behavior we can rely on the fact that this internal logic sets a timer before an actual editing begins.

ListView provides a standard way to cancel a pending label edit - `LVM_CANCELEDITLABEL` message. The problem is that this message is only implemented in 6.0 version of ComCtl32.dll. It also doesn't work as expected under 6.0, so it can't be actually used.

Another way is to cancel the timer using `KillTimer`. The problem with this approach is that the timer can be already elapsed at the moment of handling by our code (the message is already on the queue, so `KillTimer` won't help) because of the OS scheduler. Another problem with this approach is that it is not known what side effects may happen when the timer is cancelled by the third-party code.

We can also invoke the timer by pushing `WM_TIMER` message to the queue. This will start label editing immediately, but we need a way to block it. One way is to temporarily change the LabelEdit property - disable it before `WM_TIMER` is sent and enable it after the timer was handled. This works, but may cause redrawing. Another way is to handle `LVN_BEGINLABELEDITW` message which allows to cancel label editing. This is actually the approach which is implemented here, we cancel label editing only when we're sending `WM_TIMER` message. This way we don't have a potential redrawing issue.