Fix Incorrect rendering of button borders #6187
The issue is reproducible because in the original fix we were drawing a rectangle over the border of the button. Unfortunately, Windows 11 uses standard buttons with rounded edges, so the previous fix is not suitable.

As a fix, logic was added that changes the color of the border using "CompatibleDC". This saves us from having to draw different borders for Windows 10 and Windows 11. In addition, we do not have special methods for drawing a rectangle with rounded edges.

Added a special "GetRequiredColor" method for working with contrast and getting the required contrast ratio by darkening or lightening the color.