Fixed `RuntimeId` for system accessible wrapper in AccessibleObject (#6192)
Fixes #6165. `AccessibleObject.RuntimeId` was previously changed to assert and throw an excepton to avoid subclasses with non-overriden `RuntimeId` property (see #5638 for details). But it was overlooked that `AccessibleObject` is also used to wrap system accessible objects. Because of that you may get a crash or hit an assertion (with debugger attached) when running a debug version of the code. This doesn't affect users of WinForms since the SDK is compiled in release mode.

To fix the issue a separate subclass of `AccessibleObject` may be created. But this will probably be a big and complex change. Instead of that we just can add a branch for system wrapper to `RuntimeId` property.

An empty array can be used as a `RuntimeId` value for system wrappers. This will cause the UIAutomationCore to perform field-by-field comparison instead of `RuntimeId` comparison. It doesn't look as a good option since field-by-field comparison may be bad for performance and also it will be better to not rely on the UIAutomationCore internal behavior.

Hash code value was applied as a valid `RuntimeId` for system wrappers. It is better than an empty array since it doesn't cause any performance penalties and is explicit so we don't rely on UIAutomationCore behavior.