Fixed announcing the tooltip for the TabPage (#6489)
Fixes #5835. There are some controls that are adding their own tools (e.g. TabControls adds a tool for each TabPage) and for these tools we don't have an associated caption. Moreover these tools are typically retrieving the caption using TTN_NEEDTEXT notification.

The simplest way to fix this issue is to retrieve the caption using the `GetWindowText` function. At `WmShow` we know that the caption was already retrieved and set by the tooltip, so it can be retrieved using this API function. This way works for both static and dynamic captions.

The more complex way is to change the logic of the current tool retrieval and also the most important to change the logic of the caption retrieval, to support the dynamic ones.

The first approach was chosen as it is very simple and works without any downsides. The second approach also means that we are actually retrieving the caption twice: once by the tooltip internal logic and then once by our announcing logic.