The issue is reproduced because we have cached the ListViewItem.AccessibilityObject. After the changes in https://github.com/dotnet/winforms/pull/4910, we added logic that, with each call to the "AccessibilityObject" property, checked that the ListView exists and, if absent, threw an exception. (#6352)
We now return null as a ListViewItem.AccessibilityObject instead of returning an exception.

Also, after the focused ListViewItem is removed, it remains cached in the FocusedItem property. It also forces the Accessibility tool to try to retrieve the data of the ListViewItem that is not displayed. As a fix, a logic was added to reset the Focused flag from the ListViewItem.