Fix [Accessibility] Narrator announces NumericUpDown as "edit" instead of "spinbutton" #6359
The issue is reproduced when the NumericUpDown.AccessibleObject has no name. In this case, we simply announce the "edit" for text field.

This behavior can be considered expected by default. But in accordance with the behavior of the combobox, I suggest announcing the default name for the editor (spinner) if the NumericUpDown.AccessibleObject  did not return a name.

Added default name for text field inside NumericUpDown and DomainUpDown.
Added unit tests