Fixing Popup button border color
Fixes #6556. which is reproduced because in the case described in the repro steps, an additional border is drawn by the "DrawDefaultBorder" method. Since they use a different color, the button comes with a two-color border. (#6557)
Fixes #6560 is reproduced because the original color has low contrast.

Added logic for color darkening. Changed the logic for drawing borders so that the same color is used for both methods as it was before.