Fixes #6259.
When moving the Form to a new monitor with different DPI settings, Windows changing the Dpi
on the handle and sendign DPI changed message to the Form are not in sync. Due to this,
method User32.GetClientRect are returnign the incorrect value.  This method is not
accurate with respect to adornments metrics and their Dpi. Hence this work around.

This method is expected to be called whenever there is WINDOWSPOSITION_CHANGED message is triggered.
We will be calling UpdateBounds(x, y, width, height) explicitly, instead, for MDI child windows
if we had to chnage the size. for every other control, bounds are changed here.