Fixing PGV accessibility tree for property descriptors and dropdown holders (#6341)
* Fixed PGV accessibility tree for property descriptors and dropdown holders

Resolves #2848 by fixing sibiling relationship between child property descriptors and selected entry controls. Also resolves #4876 by fixing parent-child relationship between drop-down holder and listbox.

The fixes are implemented by adding helper methods to retrieve a sibling node for a specified child node. This implementation removes the need for copy-paste throughout different child nodes of property descriptor entry, it greatly simplifies the implementation and helps to avoid mistakes in future because the implementation is now located in a single place.