Fixing Anchor Info calculations. (#6778)
* Fixing AnchorInfo calculations.

Fixes #5774
Updating AnchoriInfo is only needed when control is ready for layout. InitLayoutCore() does call UpdateAnchorInfo().
At the least, we are checking if control is parented before updating AnchorInfo. This helps avoid calculating
AnchorInfo with default initial values of the Control. They are always overriden when layout happens.

* Update the condition.

* Adding integration test for Ancho and Dock layouts.

* Fix compile error.

* Adding header to test files.

* Review feedback

* Fixing AnchorInfo calculations.

Fixes #5774
Updating AnchoriInfo is only needed when control is ready for layout. InitLayoutCore() does call UpdateAnchorInfo().
At the least, we are checking if control is parented before updating AnchorInfo. This helps avoid calculating
AnchorInfo with default initial values of the Control. They are always overriden when layout happens.

* Adding integration test for Ancho and Dock layouts.

* Adding header to test files.

* Merge compile fix.