Reworked the base TextPattern implementation.
Moved the real implementation into derived types (TextBoxBaseAccessibleObject and ComboBoxChildEditUiaProvider).
Reduced the logic of edit filed that have TextPattern behavior, inherit them from TextBoxBaseAccessibleObject, thereby take many members from it and reduce the size of the types, make them simpler.
The main reason of this change - 95% of accessible objects don't implement TextPattern and they don't know about this pattern and shouldn't keep text providers that are not used. The decision is based on discussions of CF #377144 (Bug 1043396)

PS. ToolStripTextBoxControlAccessibleObject don't need to inherit FragmentNavigate and FragmentRoot from ToolStripHostedControlAccessibleObject because the base ControlAccessibleObject class do it for any ControlHost items.