Reverse resetting the Anchorinfo when size chnages.
This was regression caused by #6778 that was helping optimize anchor info calculation.
However, in certain scenarios, We access AnchorInfo by assuming it was initialized at the time of handle creation and resetting with size
change is causing null reference exception.  Given regression was introduced in the preview,
taking this fix while we are investigating on how to get anchorlayout working for
higher DPI machines.

Fixes #7159 and #7090