Implement a fix for #3161
Currently the MsgBox and InputBox APIs showed Microsoft.VisualBasic.Core (or other assembly based on inlining), instead of the calling assembly's name.
This commit fixes the regression so that it uses the calling assembly's name, or product name if specified in the project file (of the calling assembly). Or it could use more fallbacks if those fail (see code).
The calling assembly could still be inlined, but there is nothing we can do about that here and it is up to the library to add NoInlining if they feel the need to guarantee the name and not just use the Title parameter.
It also changes (in InputBox) the `String.IsNullOrEmpty(Title)` check to `Title Is Nothing` for consistency with the MsgBox API.