Ignore tooltip null-owners
Resovlves #7262

In some circumstances it's plausible to have a race condition, and the
control, for which a tooltip is being created, may have disappeared by
the time the tooltip is about to be shown. This may lead to the following:

System.NullReferenceException: Object reference not set to an instance of an object.
   at System.Windows.Forms.Control.GetSafeHandle(IWin32Window window)
   at Interop.ComCtl32.ToolInfoWrapper`1..ctor(T handle, TTF flags, String text)
   at System.Windows.Forms.ToolTip.TryGetBubbleSize(IKeyboardToolTip tool, Rectangle toolRectangle, Size& bubbleSize)
   at System.Windows.Forms.ToolTip.ShowKeyboardToolTip(String text, IKeyboardToolTip tool, Int32 duration)
   at System.Windows.Forms.KeyboardToolTipStateMachine.ShowToolTip(IKeyboardToolTip tool, ToolTip toolTip)

Ignore such situations.