Reworked UpDown controls accessible objects implementations.
Made code refactoring, moved the implementation into the base class.
Based on the approach of #4207 - "The Items list of the DomainUpDown control couldn't be visible at UI ... So, not showing the "Items" list in the Inspect tool is a new, but expected behavior." So removed useless "ItemList" and "Item" accessible objects. They are never created and used.
This change is a preparation of fixing of #7328.