Fixing issue with announcing of PropertyGrid expanded state (#7242)
I think the problem is being reproduced due to competitive requests to Narrator. Periodically it announces expanded state information and periodically skips.

The fix is to replace the standard method "RaiseAutomationPropertyChangedEvent" with the "RaiseAutomationNotification" method. In this case, we force the expanded state information to be announced, which helps avoid a possible contention problem.

Co-authored-by: SergeySmirnov-Akvelon <sergey.smirnov@akvelon.com>