Reworked MonthCalendar accessibility tree rebuilding process after visible range changing, including view changing.
Removed extra methods of collection cleaning, because DisconnectChildren methods do it as well. So avoid double cleaning.

Fixes #7714

The reason of the bug was calling RebuildAccessibilityTree 2 times at the same time via CalendarViewChanged and DisplayRangeChanged events.
Removed CalendarViewChanged event using because it duplicates the work of DisplayRangeChanged and leads to the exception. We release elements the first rebuilding time and cath the exception while the second call.