Fix Interop.Gdi32.StartDoc p/invoke (dotnet/runtime#76569)
StartDoc takes a pointer to a DOCINFO struct. We had DOCINFO defined as a class on the managed side such that built-in marshalling would marshal it as a pointer. With the switch to custom marshalling via NativeMarshalling, we ended up marshalling the struct that is the native representation rather than a pointer to it.

This switches the DOCINFO to a struct and updates the parameter to StartDoc to be in, such that it should be marshalled as a pointer to the struct by both generated and built-in marshalling.

Commit migrated from https://github.com/dotnet/runtime/commit/a6d396046ab327281e5a436b8d525f0b51d2ce0c