Convert file dialog COM usage to CsWin32
Directly uses COM pointers when accessing the native dialog.

This also rewrites the CCW for the dialog options to use the CsWin32 generated interface and moves the statics into the CsWin32 COM struct (which is partial).

Adds better helpers to initialize vtables and find the CCW for a given object.

Removes our DECIMAL and BSTR definitions.

Adds an enum helper for setting/clearing flags based on a boolean value.