Reorder Message fields
Historically our first four fields matched the layout of MSG. Keeping it that way for better back compat.

As Message doesn't match MSG, they shouldn't be blitted over each other, but for those that might be doing this we can keep the order. Note that other future changes to the struct could break doing this sort of thing even if the first four fields stay the same.