Fix Anchors in PermonitorV2 mode applications. (#8086)
* Fix Anchors in PermonitorV2 mode applications.

In PermonitorV2 mode applications, when moving the Form from one monitor to the other, Form gets an invalid display rectangle for a brief period during the scaling of its child controls. Updating anchors, because of layout, during this period would end up invalid
anchor calculations. For all scaling scenarios due to DPI changed event, we will be scaling anchors and avoid/skip recalculation of anchors and prevent using this invalid display rect.

Following is the snippet where we see invalid DisplayRect of the Form for a brief period while moving from one monitor to the other.

* Spell check.