Add support for the COM global interface table
When we're holding onto COM pointers that need to be released in a finalizer we need to ensure that we're hitting an agile reference to get any needed apartment marshalling. The Global Interface Table provides a mechanism for this down to our lowest supported platforms.

This creates a static wrapper for the GIT and a disposable wrapper class `AgileComPointer` for managing access to the GIT.

This moves AxHost.State to utilize the new `AgileComPointer`.

Also add a global using for XUnit for the interop unit tests.