Fix COM TypeDescriptor SetValue (#8246)
The main purpose of the change was to start adding type descriptor tests for our COM descriptor (which is used by System.ComponentModel). Discovered that SetValue was broken, so the title is such.

This adds a starting set of tests, refactors IPicture/Disp helpers into a shared location, and references the outputs from AxHosts.