Fix double sclaing for scrollbars
Fixes #8270

in .NET 6, we streamlined scaling of child controls to use `AutScaleFactor` in PMv2 mode applications.
 This is a chnage from .NET framework, where we were scaling child controls with DPI factor even when
`AutoScaleMode` was `Font`. As a result, scaling is centralized under `OnParentFontChanged` event and
any controls that were doind explicit scaling become redundant.

When it comes to `ScrollBars`, we added public property in .NET Framework 4.8 to let developers opt-in into
DPI scaling for scroll bars. We are preserving this public property to contorl DPI scaling.