Cache Top-level Window Form's size in PMV2 mode applications (#8333)
* Cache Top-level Window Form's size in PMV2 mode applications

When `AutoScaleMode` is `Font`, for PMV2 mode applications, when top-level `Form` is moved from one monitor to the other with different DPI settings, we compute the new size based on the scale factor we compute for the `Font` on the new monitor. This sometimes, especially when the DPI setting on those monitors is not multiple of 100s, causes rounding issues in the size computations. Rounding issue gets worse when user move top-level `Form` between monitors multiple times. Caching size for the Form, prevents rounding issues in this scenario.

* Add tests and fix a regression.

* Address feedback.